atualizar_bd_contadoria/requirements.txt
pandas
supabase
openpyxl
python-dotenv
